Take and Drop initial functions added
Player has 2 new functionalities to take and drop and item. Intentory also implemented with max size of 3. Once and item is taken, it disappears form the room, and when an item is dropped it appears in the room, so now items are exchangable from the rooms.

Also created Utils which I might expand later if needed, but for now it has a function to control case sensitive which I used at picking and dropping item time.