Final touch added
A bit of history and side quest added to the world. Now there is the option for fight one of the npc with a sword that will drop an item when killed. If you return this item to the old man, it will change it's default dialog and thank you for returning it to him.

Attack command added to player.
Look command added to player.
More things added to world.